:recycle: Refactor: Folder 페이지 컴포넌트 모듈화 작업
