--- src/components/Folder/FolderHeader.js
@@ -1,14 +1,0 @@
-import './Folder.css';
-import AddLinkBar from '../AddLinkBar/AddLinkBar';
-
-function FolderHeader() {
-  return (
-    <nav className="nav">
-      <div className="nav-item">
-        <AddLinkBar />
-      </div>
-    </nav>
-  );
-}
-
-export default FolderHeader;
